requirements.txt
aiohttp==3.8.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.9.2
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
asyncssh==2.13.0
atpublic==3.1.1
attrs==22.2.0
billiard==3.6.4.0
boto3==1.26.54
botocore==1.29.54
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cligj==0.7.2
cloudpickle==2.2.1
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.8
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
databricks-cli==0.17.4
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.8.0
docker==6.0.1
dpath==2.1.4
dulwich==0.21.2
dvc==2.41.1
dvc-data==0.29.0
dvc-http==2.30.0
dvc-objects==0.14.1
dvc-render==0.0.17
dvc-studio-client==0.3.0
dvc-task==0.1.9
dvclive==1.3.2
entrypoints==0.4
fastapi==0.89.1
filelock==3.9.0
Fiona==1.8.22
Flask==2.2.2
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
funcy==1.17
future==0.18.3
geopandas==0.12.2
gitdb==4.0.10
GitPython==3.1.30
grandalf==0.6
greenlet==2.0.1
h11==0.14.0
hydra-core==1.3.1
idna==3.4
importlib-metadata==5.2.0
iterative-telemetry==0.0.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
lightgbm==3.3.4
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
mlflow==2.1.1
multidict==6.0.4
munch==2.5.0
nanotime==0.5.2
networkx==3.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
packaging==22.0
pandas==1.5.3
pathspec==0.9.0
Pillow==9.4.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
pyarrow==10.0.1
pycparser==2.21
pydantic==1.10.4
pydot==1.4.2
pygit2==1.11.1
Pygments==2.14.0
pygtrie==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
pyproj==3.4.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pywin32==305
PyYAML==6.0
querystring-parser==1.2.4
requests==2.28.2
rich==13.1.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
scikit-learn==1.2.0
scipy==1.10.0
scmrepo==0.1.5
shap==0.41.0
shapely==2.0.0
shortuuid==1.0.11
shtab==1.5.8
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
starlette==0.22.0
tabulate==0.9.0
threadpoolctl==3.1.0
tomlkit==0.11.6
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
vine==5.0.0
voluptuous==0.13.1
waitress==2.1.2
wcwidth==0.2.6
websocket-client==1.4.2
Werkzeug==2.2.2
yarl==1.8.2
zc.lockfile==2.0
zipp==3.11.0
